Improve code quality: refactor PersistentState, split menu_builder, fix clippy/unsafe issues
- Refactor PersistentState to extract PerTypeSettings struct, reducing struct_excessive_bools
- Add PersistentStateFlat serde shadow struct for backward-compatible JSON serialization
- Split menu_builder/mod.rs (611 LOC) into device_section.rs and priority_section.rs
- Fix SAFETY comments on unsafe blocks to be directly adjacent to unsafe keyword
- Add missing SAFETY comment for GetCount call
- Fix semicolon_if_nothing_returned in SetMasterVolumeLevelScalar unsafe block
- Replace unwrap() with expect() in test code, add allow(clippy::expect_used) for tests
- Remove unused struct_excessive_bools Cargo.toml allow (moved to local #[allow])
- Skip inline_always and float_cmp clippy warnings from macro-generated/std library code
