+++ Cargo.toml
@@ -63,7 +63,7 @@ winresource = "0.1.31"
 [lints.clippy]
 # Generated by #[implement] macro — cannot be fixed without the allow
 inline_always = "allow"
-# PersistentState has 4+ boolean config flags; this is inherent to config structs
+# Serde shadow struct (PersistentStateFlat) preserves flat JSON format with per-device-type bool fields
 struct_excessive_bools = "allow"
 # All f32 comparisons in tests are exact-value checks on clamped constants (0.0, 50.0, 100.0)
 float_cmp = "allow"
