+++ src/audio/mod.rs
@@ -378,10 +378,8 @@ pub(crate) mod tests {
         // Make dev1 fail on lookup — enforce should skip it and not switch
         backend.failing_device_ids.borrow_mut().push("dev1".into());
 
-        let state = PersistentState {
-            output_priority_list: vec!["dev1".into(), "dev2".into()],
-            ..Default::default()
-        };
+        let mut state = PersistentState::default();
+        state.output.priority_list = vec!["dev1".into(), "dev2".into()];
         let mut throttler = NotificationThrottler::new();
         let temp = TemporaryPriorities {
             output: None,
@@ -408,10 +406,8 @@ pub(crate) mod tests {
         backend.set_default("dev3", DeviceType::Output);
         backend.failing_device_ids.borrow_mut().push("dev1".into());
 
-        let state = PersistentState {
-            output_priority_list: vec!["dev1".into(), "dev2".into()],
-            ..Default::default()
-        };
+        let mut state = PersistentState::default();
+        state.output.priority_list = vec!["dev1".into(), "dev2".into()];
         let mut throttler = NotificationThrottler::new();
         let temp = TemporaryPriorities {
             output: None,
@@ -436,10 +432,8 @@ pub(crate) mod tests {
         backend.set_default("dev2", DeviceType::Output);
         *backend.set_default_fails.borrow_mut() = true;
 
-        let state = PersistentState {
-            output_priority_list: vec!["dev1".into(), "dev2".into()],
-            ..Default::default()
-        };
+        let mut state = PersistentState::default();
+        state.output.priority_list = vec!["dev1".into(), "dev2".into()];
         let mut throttler = NotificationThrottler::new();
         let temp = TemporaryPriorities {
             output: None,
@@ -460,10 +454,8 @@ pub(crate) mod tests {
         ]);
         backend.set_default("other", DeviceType::Output);
 
-        let mut state = PersistentState {
-            output_priority_list: vec![DeviceId::from("old_id")],
-            ..Default::default()
-        };
+        let mut state = PersistentState::default();
+        state.output.priority_list = vec![DeviceId::from("old_id")];
         state.devices.insert(
             DeviceId::from("old_id"),
             make_device_settings("Speakers", DeviceType::Output),
@@ -472,7 +464,7 @@ pub(crate) mod tests {
         migrate_device_ids(&backend, &mut state);
         assert!(state.devices.contains_key("new_id"));
         assert!(!state.devices.contains_key("old_id"));
-        assert_eq!(state.output_priority_list, vec!["new_id"]);
+        assert_eq!(state.output.priority_list, vec!["new_id"]);
 
         let mut throttler = NotificationThrottler::new();
         let temp_priorities = TemporaryPriorities {
