+++ src/config/mod.rs
@@ -7,80 +7,115 @@ use crate::types::{DeviceId, DeviceType};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-#[derive(Debug, Serialize, Deserialize)]
+/// Per-device-type preferences (one instance for output, one for input).
+#[derive(Debug, Clone, Default)]
+pub(crate) struct PerTypeSettings {
+    pub priority_list: Vec<DeviceId>,
+    pub notify_on_priority_restore: bool,
+    pub switch_communication_device: bool,
+}
+
+/// Flat serde representation for backward-compatible JSON serialization.
+#[derive(Serialize, Deserialize)]
 #[serde(default)]
+struct PersistentStateFlat {
+    devices: HashMap<DeviceId, DeviceSettings>,
+    output_priority_list: Vec<DeviceId>,
+    input_priority_list: Vec<DeviceId>,
+    notify_on_priority_restore_output: bool,
+    notify_on_priority_restore_input: bool,
+    switch_communication_device_output: bool,
+    switch_communication_device_input: bool,
+    check_updates_on_launch: bool,
+}
+
+impl Default for PersistentStateFlat {
+    fn default() -> Self {
+        let state = PersistentState::default();
+        state.into()
+    }
+}
+
+impl From<PersistentStateFlat> for PersistentState {
+    fn from(flat: PersistentStateFlat) -> Self {
+        Self {
+            devices: flat.devices,
+            output: PerTypeSettings {
+                priority_list: flat.output_priority_list,
+                notify_on_priority_restore: flat.notify_on_priority_restore_output,
+                switch_communication_device: flat.switch_communication_device_output,
+            },
+            input: PerTypeSettings {
+                priority_list: flat.input_priority_list,
+                notify_on_priority_restore: flat.notify_on_priority_restore_input,
+                switch_communication_device: flat.switch_communication_device_input,
+            },
+            check_updates_on_launch: flat.check_updates_on_launch,
+        }
+    }
+}
+
+impl From<PersistentState> for PersistentStateFlat {
+    fn from(state: PersistentState) -> Self {
+        Self {
+            devices: state.devices,
+            output_priority_list: state.output.priority_list,
+            input_priority_list: state.input.priority_list,
+            notify_on_priority_restore_output: state.output.notify_on_priority_restore,
+            notify_on_priority_restore_input: state.input.notify_on_priority_restore,
+            switch_communication_device_output: state.output.switch_communication_device,
+            switch_communication_device_input: state.input.switch_communication_device,
+            check_updates_on_launch: state.check_updates_on_launch,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(from = "PersistentStateFlat", into = "PersistentStateFlat")]
 pub struct PersistentState {
     pub(crate) devices: HashMap<DeviceId, DeviceSettings>,
-    #[serde(rename = "output_priority_list")]
-    pub(crate) output_priority_list: Vec<DeviceId>,
-    #[serde(rename = "input_priority_list")]
-    pub(crate) input_priority_list: Vec<DeviceId>,
-    #[serde(rename = "notify_on_priority_restore_output")]
-    pub(crate) notify_on_priority_restore_output: bool,
-    #[serde(rename = "notify_on_priority_restore_input")]
-    pub(crate) notify_on_priority_restore_input: bool,
-    #[serde(rename = "switch_communication_device_output")]
-    pub(crate) switch_communication_device_output: bool,
-    #[serde(rename = "switch_communication_device_input")]
-    pub(crate) switch_communication_device_input: bool,
+    pub(crate) output: PerTypeSettings,
+    pub(crate) input: PerTypeSettings,
     pub check_updates_on_launch: bool,
 }
 
 impl PersistentState {
-    /// Single dispatch point for all per-device-type fields (shared ref).
-    fn fields_ref(&self, dt: DeviceType) -> (&Vec<DeviceId>, bool, bool) {
+    fn per_type(&self, dt: DeviceType) -> &PerTypeSettings {
         match dt {
-            DeviceType::Output => (
-                &self.output_priority_list,
-                self.notify_on_priority_restore_output,
-                self.switch_communication_device_output,
-            ),
-            DeviceType::Input => (
-                &self.input_priority_list,
-                self.notify_on_priority_restore_input,
-                self.switch_communication_device_input,
-            ),
+            DeviceType::Output => &self.output,
+            DeviceType::Input => &self.input,
         }
     }
 
-    /// Single dispatch point for all per-device-type fields (mutable ref).
-    fn fields_mut(&mut self, dt: DeviceType) -> (&mut Vec<DeviceId>, &mut bool, &mut bool) {
+    fn per_type_mut(&mut self, dt: DeviceType) -> &mut PerTypeSettings {
         match dt {
-            DeviceType::Output => (
-                &mut self.output_priority_list,
-                &mut self.notify_on_priority_restore_output,
-                &mut self.switch_communication_device_output,
-            ),
-            DeviceType::Input => (
-                &mut self.input_priority_list,
-                &mut self.notify_on_priority_restore_input,
-                &mut self.switch_communication_device_input,
-            ),
+            DeviceType::Output => &mut self.output,
+            DeviceType::Input => &mut self.input,
         }
     }
 
     pub fn priority_list(&self, device_type: DeviceType) -> &[DeviceId] {
-        self.fields_ref(device_type).0
+        &self.per_type(device_type).priority_list
     }
 
     pub fn priority_list_mut(&mut self, device_type: DeviceType) -> &mut Vec<DeviceId> {
-        self.fields_mut(device_type).0
+        &mut self.per_type_mut(device_type).priority_list
     }
 
     pub fn notify_on_priority_restore(&self, device_type: DeviceType) -> bool {
-        self.fields_ref(device_type).1
+        self.per_type(device_type).notify_on_priority_restore
     }
 
     pub fn set_notify_on_priority_restore(&mut self, device_type: DeviceType, value: bool) {
-        *self.fields_mut(device_type).1 = value;
+        self.per_type_mut(device_type).notify_on_priority_restore = value;
     }
 
     pub fn switch_communication_device(&self, device_type: DeviceType) -> bool {
-        self.fields_ref(device_type).2
+        self.per_type(device_type).switch_communication_device
     }
 
     pub fn set_switch_communication_device(&mut self, device_type: DeviceType, value: bool) {
-        *self.fields_mut(device_type).2 = value;
+        self.per_type_mut(device_type).switch_communication_device = value;
     }
 
     pub fn device_settings(&self, device_id: &DeviceId) -> Option<&DeviceSettings> {
@@ -136,8 +171,8 @@ impl PersistentState {
         if !is_prunable {
             return;
         }
-        let in_priority = self.output_priority_list.contains(device_id)
-            || self.input_priority_list.contains(device_id);
+        let in_priority = self.output.priority_list.contains(device_id)
+            || self.input.priority_list.contains(device_id);
         if !in_priority {
             self.devices.remove(device_id);
         }
@@ -148,12 +183,14 @@ impl Default for PersistentState {
     fn default() -> Self {
         Self {
             devices: HashMap::default(),
-            output_priority_list: Vec::default(),
-            input_priority_list: Vec::default(),
-            notify_on_priority_restore_output: false,
-            notify_on_priority_restore_input: false,
-            switch_communication_device_output: true,
-            switch_communication_device_input: true,
+            output: PerTypeSettings {
+                switch_communication_device: true,
+                ..PerTypeSettings::default()
+            },
+            input: PerTypeSettings {
+                switch_communication_device: true,
+                ..PerTypeSettings::default()
+            },
             check_updates_on_launch: true,
         }
     }
@@ -171,19 +208,22 @@ mod tests {
     fn persistent_state_default_values() {
         let state = PersistentState::default();
         assert!(state.devices.is_empty());
-        assert!(state.output_priority_list.is_empty());
-        assert!(state.input_priority_list.is_empty());
-        assert!(!state.notify_on_priority_restore_output);
-        assert!(!state.notify_on_priority_restore_input);
-        assert!(state.switch_communication_device_output);
-        assert!(state.switch_communication_device_input);
+        assert!(state.output.priority_list.is_empty());
+        assert!(state.input.priority_list.is_empty());
+        assert!(!state.output.notify_on_priority_restore);
+        assert!(!state.input.notify_on_priority_restore);
+        assert!(state.output.switch_communication_device);
+        assert!(state.input.switch_communication_device);
         assert!(state.check_updates_on_launch);
     }
 
     #[test]
     fn persistent_state_serialization_roundtrip() {
         let state = PersistentState {
-            output_priority_list: vec!["device_a".into(), "device_b".into()],
+            output: PerTypeSettings {
+                priority_list: vec!["device_a".into(), "device_b".into()],
+                ..PerTypeSettings::default()
+            },
             check_updates_on_launch: false,
             devices: HashMap::from([(
                 "test_id".into(),
@@ -204,7 +244,7 @@ mod tests {
         let json = serde_json::to_string_pretty(&state).unwrap();
         let loaded: PersistentState = serde_json::from_str(&json).unwrap();
 
-        assert_eq!(loaded.output_priority_list, vec!["device_a", "device_b"]);
+        assert_eq!(loaded.output.priority_list, vec!["device_a", "device_b"]);
         assert!(!loaded.check_updates_on_launch);
         let dev = loaded.devices.get("test_id").unwrap();
         assert!(dev.volume_lock.is_locked);
@@ -216,15 +256,21 @@ mod tests {
     fn persistent_state_deserialize_missing_fields_uses_defaults() {
         let json = r#"{"devices": {}}"#;
         let state: PersistentState = serde_json::from_str(json).unwrap();
-        assert!(state.output_priority_list.is_empty());
+        assert!(state.output.priority_list.is_empty());
         assert!(state.check_updates_on_launch);
     }
 
     #[test]
     fn get_priority_list_returns_correct_type() {
         let state = PersistentState {
-            output_priority_list: vec!["out1".into()],
-            input_priority_list: vec!["in1".into(), "in2".into()],
+            output: PerTypeSettings {
+                priority_list: vec!["out1".into()],
+                ..PerTypeSettings::default()
+            },
+            input: PerTypeSettings {
+                priority_list: vec!["in1".into(), "in2".into()],
+                ..PerTypeSettings::default()
+            },
             ..Default::default()
         };
 
@@ -242,8 +288,8 @@ mod tests {
             .priority_list_mut(DeviceType::Input)
             .push("new_in".into());
 
-        assert_eq!(state.output_priority_list, vec!["new_out"]);
-        assert_eq!(state.input_priority_list, vec!["new_in"]);
+        assert_eq!(state.output.priority_list, vec!["new_out"]);
+        assert_eq!(state.input.priority_list, vec!["new_in"]);
     }
 
     #[test]
@@ -272,10 +318,16 @@ mod tests {
 
         // Build a non-trivial state
         let state = PersistentState {
-            output_priority_list: vec!["dev_a".into(), "dev_b".into()],
-            input_priority_list: vec!["mic_1".into()],
-            notify_on_priority_restore_output: true,
-            switch_communication_device_input: false,
+            output: PerTypeSettings {
+                priority_list: vec!["dev_a".into(), "dev_b".into()],
+                notify_on_priority_restore: true,
+                ..PerTypeSettings::default()
+            },
+            input: PerTypeSettings {
+                priority_list: vec!["mic_1".into()],
+                switch_communication_device: false,
+                ..PerTypeSettings::default()
+            },
             check_updates_on_launch: false,
             devices: HashMap::from([(
                 "dev_a".into(),
@@ -293,7 +345,6 @@ mod tests {
                     name: "Speakers".into(),
                 },
             )]),
-            ..Default::default()
         };
 
         // Write to file
@@ -304,10 +355,10 @@ mod tests {
         let data = fs::read_to_string(&path).unwrap();
         let loaded: PersistentState = serde_json::from_str(&data).unwrap();
 
-        assert_eq!(loaded.output_priority_list, vec!["dev_a", "dev_b"]);
-        assert_eq!(loaded.input_priority_list, vec!["mic_1"]);
-        assert!(loaded.notify_on_priority_restore_output);
-        assert!(!loaded.switch_communication_device_input);
+        assert_eq!(loaded.output.priority_list, vec!["dev_a", "dev_b"]);
+        assert_eq!(loaded.input.priority_list, vec!["mic_1"]);
+        assert!(loaded.output.notify_on_priority_restore);
+        assert!(!loaded.input.switch_communication_device);
         assert!(!loaded.check_updates_on_launch);
 
         let dev = loaded.devices.get("dev_a").unwrap();
@@ -358,7 +409,7 @@ mod tests {
             .unwrap()
             .volume_lock
             .target_percent = VolumePercent::from(75.0);
-        loaded.output_priority_list.push("dev1".into());
+        loaded.output.priority_list.push("dev1".into());
 
         let json2 = serde_json::to_string_pretty(&loaded).unwrap();
         fs::write(&path, &json2).unwrap();
@@ -369,7 +420,7 @@ mod tests {
         let dev = final_state.devices.get("dev1").unwrap();
         assert!(dev.volume_lock.is_locked);
         assert_eq!(dev.volume_lock.target_percent, 75.0);
-        assert_eq!(final_state.output_priority_list, vec!["dev1"]);
+        assert_eq!(final_state.output.priority_list, vec!["dev1"]);
 
         // Cleanup
         let _ = fs::remove_file(&path);
